New unconnected blocks will automatically connect to the bottom container/statement
If a block cannot stand alone, then the user will be alerted audibly
that the block needs to be connected. Statement blocks(if, while etc)
and function blocks will stand alone. Because functions cannot be
connected to statements, the function will be placed directly below the
last statement. If another statement is added afterwards, it will be
placed below the function.